build: add Node 16 and 17 to the CI matrix
Update documentation, so that it does not become outdated whenever a new version of Node is released.

Test on the current version of Node, so we can spot problems early, but don't claim to support it as current release line often introduces bugs, which are later fixed by Node itself.

Fixes #3730, fixes #3728
Closes #3729